Refactor to inner control API
Join link and awkard methods to a control API.

The main module stores the complete key and offers the control API
for the key to the "node-form" module which is focused on
providing single node editing, creation and navigation.

Draw the node label in white if the contrast becomes too bad.